cdk/apps/htsget/lambdas/requirements.txt
PyJWT==2.1.0
cryptography==3.4.7
requests==2.25.1
